feat(py)!: no default command for verify
Command is picked up from sea.yaml `command` property.
Due to bug in SeaHorn, for older version this means that command
must be specified on command line. For example,

./verify --command=fpf REST

instead of

./verify REST

The default command was `fpf` so the above transformation should work for most users.

New version of SeaHorn correctly pick up the command from the yaml file.